Tile Hover, Work Area Sizing, and Pawn Count
Added feature where when a user hovers over a tile, info on that tile is shown (will need to update this once we incorporate whatever version of PlacedObjects). Also added a UI element to show the size of a Work Area when a user is drag and dropping it. Finally, hooked up LaborOrderManager's "PawnCount()" to the Colony Info UI so the correct pawn count is now shown.